robertalab/#464 added RedHat-based Linux functionality (tested on Fedora25)

+++ Resources/USBInstaller/Linux/install.sh
@@ -2,9 +2,15 @@
 
 dir="$( pwd )"
 
-dpkg --add-architecture i386
-apt-get update
-apt-get install -y libc6-i386 libxext6:i386 libxrender1:i386 libxtst6:i386 libxi6:i386 libusb-0.1-4:i386
+if [ -f /etc/debian_version ]; then
+    dpkg --add-architecture i386
+    apt-get update
+    apt-get install -y libc6-i386 libxext6:i386 libxrender1:i386 libxtst6:i386 libxi6:i386 libusb-0.1-4:i386
+elif [ -f /etc/redhat-release ]; then
+    dnf update
+    dnf install glibc.i686 libXext.i686 libXrender.i686 libXtst.i686 libXi.i686 libusb.i686 
+#elif ... other distros
+fi
 
 
 file="/etc/udev/rules.d/70-lego.rules"
@@ -15,7 +21,7 @@ else
   echo "# Lego NXT brick in normal mode" >> "/etc/udev/rules.d/70-lego.rules"
   echo 'SUBSYSTEM=="usb", DRIVER=="usb", ATTRS{idVendor}=="0694", ATTRS{idProduct}=="0002", GROUP="lego", MODE="0660"' >> "/etc/udev/rules.d/70-lego.rules"
   echo '# Lego NXT brick in firmware update mode (Atmel SAM-BA mode)' >> "/etc/udev/rules.d/70-lego.rules"
-  echo 'SUBSYSTEM=="usb", DRIVER=="usb", ATTRS{idVendor}=="03eb", ATTRS{idProduct}=="6124", GROUP="lego", MODE="0660"' >> "/etc/udev/rules.d/70-lego.rules"
+  echo 'SUBSYSTEM=="usb", DRIVER=="usb", ATTRS{idVendor}=="03eb", ATTRS{idProduct}=="6124", GROUP="lego", MODE="0660"' >> "/etc/udev/rules.d/70-dnflego.rules"
 fi
 
 echo "[Desktop Entry]" > "/usr/share/applications/ORUSB.desktop"
